timeline: extract world generation identity
